% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilstats.R
\name{calculate_adl_kroll_test}
\alias{calculate_adl_kroll_test}
\title{Calculate Average Deviation From Linearity Kroll Test}
\usage{
calculate_adl_kroll_test(dilution_data, conc_var, signal_var)
}
\arguments{
\item{dilution_data}{A data frame or tibble containing dilution data}

\item{conc_var}{Column name in \code{dilution_data} to indicate concentration}

\item{signal_var}{Column name in \code{dilution_data} to indicate signal}
}
\value{
The average deviation from linearity
}
\description{
Calculate the average deviation from linearity kroll test
}
\details{
The function will return NA if the number of dilution points
is less than or equal to three
}
