% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputdata-excel.R
\name{calculate_column_max_char_vector}
\alias{calculate_column_max_char_vector}
\title{Calculate Column Maximum Char}
\usage{
calculate_column_max_char_vector(dilution_summary)
}
\arguments{
\item{dilution_summary}{The summary table generated
by function \code{\link[=summarise_dilution_table]{summarise_dilution_table()}} and/or
\code{\link[=evaluate_linearity]{evaluate_linearity()}}
but it can also be any generic data frame or tibble}
}
\value{
A numeric vector each value indicated the
maximum number of characters for each column of the
input data frame or tibble
}
\description{
Calculate the maximum number of
characters for each column in the input data frame
or tibble, including the column name.
}
\examples{
r_corr <- c(0.951956, 0.948683, 0.978057, 0.976462,
            0.970618, 0.969348, 0.343838, 0.383552)
pra_linear <- c(65.78711, 64.58687, 90.21257, 89.95473,
                72.91220, 72.36528, -233.05949, -172.13659)
mandel_p_val <- c(2.899006e-07, 7.922290e-07, 2.903365e-01, 3.082930e-01,
                  3.195779e-08, 6.366588e-08, 3.634004e-02, 1.864090e-02)
concavity <- c(-4133.501328, -4146.745747, -3.350942, -3.393617,
               0.3942824, 0.4012963, -19.9469621, -22.6144875)
dilution_summary <- data.frame(r_corr = r_corr, pra_linear = pra_linear,
                               mandel_p_val = mandel_p_val,
                               concavity = concavity)
column_max_char <- calculate_column_max_char_vector(dilution_summary)
}
