% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilstats.R
\name{calculate_concavity}
\alias{calculate_concavity}
\title{Calculate Concavity}
\usage{
calculate_concavity(curve_data, conc_var, signal_var)
}
\arguments{
\item{curve_data}{A data frame or tibble containing curve data.}

\item{conc_var}{Column name in \code{curve_data} to indicate concentration.}

\item{signal_var}{Column name in \code{curve_data} to indicate signal.}
}
\value{
The concavity of the Quadratic Model.
}
\description{
Calculate the concavity of the Quadratic Model.
}
\details{
The function will return NA if the number of curve points
is less than or equal to three.
}
\examples{
concentration <- c(10, 20, 40, 60, 80, 100)

signal <- c(22561, 31178, 39981, 48390, 52171, 53410)

curve_data <- data.frame(Concentration = concentration,
                         Signal = signal)

concavity_value <- calculate_concavity(
  curve_data,
  "Concentration",
  "Signal"
)

concavity_value

}
