% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilstats.R
\name{calculate_gof_linear}
\alias{calculate_gof_linear}
\title{Calculate Linear Model's Goodness Of Fit}
\usage{
calculate_gof_linear(
  curve_data,
  conc_var,
  signal_var,
  dilution_data = lifecycle::deprecated()
)
}
\arguments{
\item{curve_data}{A data frame or tibble containing curve data.}

\item{conc_var}{Column name in \code{curve_data} to indicate concentration.}

\item{signal_var}{Column name in \code{curve_data} to indicate signal.}

\item{dilution_data}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{dilution_data} was renamed to
\code{curve_data}.}
}
\value{
A tibble containing the Goodness of Fit measures of the linear model
The Goodness of Fit measures are the Pearson correlation coefficient (R),
the R^2 and the adjusted R^2.
}
\description{
Calculate the Goodness of Fit of the Linear Model.
}
\details{
The function will return a tibble with NA values
if the number of curve points is less than or equal to three.
}
\examples{
concentration <- c(10, 20, 40, 60, 80, 100)

signal <- c(22561, 31178, 39981, 48390, 52171, 53410)

curve_data <- data.frame(Concentration = concentration,
                         Signal = signal)

curve_linear_gof <- calculate_gof_linear(
  curve_data,
  "Concentration",
  "Signal"
)

curve_linear_gof

}
