% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilstats.R
\name{calculate_pra_linear}
\alias{calculate_pra_linear}
\title{Calculate Linear Model's Percent Residual Accuracy (PRA)}
\usage{
calculate_pra_linear(
  curve_data,
  conc_var,
  signal_var,
  dilution_data = lifecycle::deprecated()
)
}
\arguments{
\item{curve_data}{A data frame or tibble containing curve data.}

\item{conc_var}{Column name in \code{curve_data} to indicate concentration.}

\item{signal_var}{Column name in \code{curve_data} to indicate signal.}

\item{dilution_data}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{dilution_data} was renamed to
\code{curve_data}.}
}
\value{
The Percent Residual Accuracy (PRA) of the
Linear Model. More information of this value can be found in
\url{https://www.sciencedirect.com/science/article/abs/pii/S0039914018307549}
}
\description{
Calculate the Percent Residual Accuracy (PRA) of the
Linear Model.
}
\details{
The function will return NA if the number of curve points
is less than or equal to three.
}
\examples{
dilution_percent <- c(10, 20, 40, 60, 80, 100)

area <- c(22561, 31178, 39981, 48390, 52171, 53410)

curve_data <- data.frame(Dilution_Percent = dilution_percent, Area = area)

pra_value <- calculate_pra_linear(curve_data, "Dilution_Percent", "Area")

pra_value

}
