% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognostics.R
\name{convert_to_cognostics}
\alias{convert_to_cognostics}
\title{Convert to cognostics}
\usage{
convert_to_cognostics(
  dilution_summary,
  cog_df = NULL,
  grouping_variable = c("Transition_Name", "Dilution_Batch"),
  col_name_vec = "col_name_vec",
  desc_vec = "desc_vec",
  type_vec = "type_vec",
  default_label_vec = "default_label_vec"
)
}
\arguments{
\item{dilution_summary}{The summary table generated
by function \code{summarise_dilution_table} and/or \code{evaluate_linearity}
but it can also be any generic data frame or tibble}

\item{cog_df}{A data frame or tibble that contains cognostics information
If no input is given the cognostics information generated by function
\code{get_dil_default_cognostics} will be used.
Default: NULL}

\item{grouping_variable}{PARAM_DESCRIPTION, Default: c("Transition_Name", "Dilution_Batch")}

\item{col_name_vec}{Column name in \code{cog_df} to indicate the columns
in \code{dilution_summary} that needs to be converted to a cognostics,
Default: 'col_name_vec'}

\item{desc_vec}{Column name in \code{cog_df} to indicate the description
for each cognostics as define in \code{trelliscopejs::cog},
Default: 'desc_vec'}

\item{type_vec}{Column name in \code{cog_df} to indicate the type
of each cognostics as define in \code{trelliscopejs::cog},
Default: 'type_vec'}

\item{default_label_vec}{Column name in \code{cog_df} to indicate the if
the given cognostics is a panel label as define in \code{trelliscopejs::cog},
Default: 'default_label_vec'}
}
\value{
The dilution summary table with \verb{groupung variable} columns
converted to conditional cognostics,
other columns in \code{dilution_summary} converted to cognostics
to be used in the in the \code{trelliscopejs} report.
}
\description{
Convert columns in \code{dilution_summary} to \code{trelliscopejs} cognostics
}
