% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilplots-utils.R
\name{create_char_seq}
\alias{create_char_seq}
\title{Create cyclic character sequence}
\usage{
create_char_seq(group_name, output_length)
}
\arguments{
\item{group_name}{A character vector as input}

\item{output_length}{The length of the output sequence}
}
\value{
A cyclic character sequence with length \code{output_length}
}
\description{
Create cyclic character sequence
}
\details{
Taken from
https://community.rstudio.com/t/fill-in-a-sequence-of-letters-based-on-a-given-order/88823/3
}
\examples{
group_name <- c("red", "green", "blue")
create_char_seq(group_name, output_length = 2)
create_char_seq(group_name, output_length = 5)

}
