% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilstats.R
\name{create_cubic_model}
\alias{create_cubic_model}
\title{Create Cubic Model}
\usage{
create_cubic_model(curve_data, conc_var, signal_var)
}
\arguments{
\item{curve_data}{A data frame or tibble containing curve data.}

\item{conc_var}{Column name in \code{curve_data} to indicate concentration.}

\item{signal_var}{Column name in \code{curve_data} to indicate signal.}
}
\value{
A linear model object from \code{stats:lm()} with formula
\code{signal_var ~ conc_var + I(conc_var * conc_var * conc_var)}
from data \code{diltuion_data}.
}
\description{
A wrapper to create a cubic model for curve data.
}
\examples{
concentration <- c(10, 20, 40, 60, 80, 100)

signal <- c(22561, 31178, 39981, 48390, 52171, 53410)

curve_data <- data.frame(Concentration = concentration,
                         Signal = signal)

cubic_model <- create_cubic_model(
  curve_data,
  "Concentration",
  "Signal"
)

cubic_model

}
