% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydata.R
\name{create_curve_table}
\alias{create_curve_table}
\title{Create Curve Table}
\usage{
create_curve_table(
  curve_annot,
  curve_data_wide,
  common_column = c("Sample_Name"),
  signal_var = "Area",
  column_group = "Transition_Name"
)
}
\arguments{
\item{curve_annot}{A data frame or tibble that contains information
of your curve batch. A column with sample name should be present.}

\item{curve_data_wide}{A wide format data frame or tibble
that contains the Sample Name usually at the first column followed
by different curves. Each curve is meant to provide
a \code{signal_var} for each sample.}

\item{common_column}{A vector consisting of common column names that
must be found in both \code{curve_annot} and \code{curve_data_wide},
Default: c("Sample_Name")}

\item{signal_var}{Value provided in \code{curve_data_wide} for each
curve which will be used as a column name when the merged data
is converted to a long format,
Default: 'Area'}

\item{column_group}{Value used to group the curves in
\code{curve_data_wide} when the merged data is converted to a long format,
Default: 'Transition_Name'}
}
\value{
A data frame or tibble in long format
}
\description{
Create Curve Table from two data frame or tibble inputs
\code{curve_annot} and \code{curve_data_wide}. Both input must have at least one
common column. The default common column names is \code{Sample_Name}.
}
\details{
We first merge the \code{curve_annot} with the \code{curve_data_wide}
via one or more common columns. Next we convert the data from a wide to
a long format. Merging with a Sample Annotation data can be done after this
}
\examples{
# Data Creation
dilution_percent <- c(
  10, 20, 40, 60, 80, 100,
  10, 20, 40, 60, 80, 100
)

dilution_batch_name <- c(
  "B1", "B1", "B1", "B1", "B1", "B1",
  "B2", "B2", "B2", "B2", "B2", "B2"
)

sample_name <- c(
  "Sample_010a", "Sample_020a", "Sample_040a",
  "Sample_060a", "Sample_080a", "Sample_100a",
  "Sample_010b", "Sample_020b", "Sample_040b",
  "Sample_060b", "Sample_080b", "Sample_100b"
)

lipid1_area <- c(
  22561, 31178, 39981, 48390, 52171, 53410,
  32561, 41178, 49981, 58390, 62171, 63410
)

lipid2_area <- c(
  2299075, 4136350, 7020062, 8922063, 9288742, 11365710,
  2300075, 4137350, 7021062, 8923063, 9289742, 11366710
)

curve_annot <- tibble::tibble(
  Sample_Name = sample_name,
  Dilution_Batch_Name = dilution_batch_name,
  Dilution_Percent = dilution_percent
)

curve_data <- tibble::tibble(
  Sample_Name = sample_name,
  Lipid1 = lipid1_area,
  Lipid2 = lipid2_area
)

# Create dilution table
curve_table <- create_curve_table(
  curve_annot = curve_annot,
  curve_data_wide = curve_data,
  common_column = "Sample_Name",
  signal_var = "Area",
  column_group = "Transition_Name"
)

print(curve_table, width = 100)

}
