% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognostics.R
\name{create_dil_default_cog_df}
\alias{create_dil_default_cog_df}
\title{Create default cognostics for dilution plot}
\usage{
create_dil_default_cog_df()
}
\value{
A dataframe of default cognostics information
}
\description{
Create default cognostics for dilution plot
}
\details{
Internal function to create a dataframe of default cognostics
information to be used by the \code{trelliscopejs::cog}.
}
