% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilstats.R
\name{create_dil_quad_model}
\alias{create_dil_quad_model}
\title{Create Dilution Quadratic Model}
\usage{
create_dil_quad_model(dilution_data, conc_var, signal_var)
}
\arguments{
\item{dilution_data}{A data frame or tibble containing dilution data}

\item{conc_var}{Column name in \code{dilution_data} to indicate concentration}

\item{signal_var}{Column name in \code{dilution_data} to indicate signal}
}
\value{
A linear model object from \code{stats:lm()} with formula
\code{signal_var ~ conc_var + I(conc_var * conc_var)}
from data \code{diltuion_data}
}
\description{
A wrapper to create a quadratic model for dilution data
}
\examples{
dilution_percent <- c(10, 20, 40, 60, 80, 100)
area <- c(22561, 31178, 39981, 48390, 52171, 53410)
dilution_data <- data.frame(Dilution_Percent = dilution_percent, Area = area)
linear_model <- create_dil_quad_model(dilution_data,
                                        "Dilution_Percent",
                                        "Area")
}
