% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydata.R
\name{create_dilution_table}
\alias{create_dilution_table}
\title{Create Dilution Table}
\usage{
create_dilution_table(
  dilution_annot,
  lipid_data_wide,
  common_column = c("Sample_Name"),
  signal_var = "Area",
  column_group = "Transition_Name"
)
}
\description{
Create Dilution Table from two data frame or tibble inputs
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{create_dilution_table} was renamed to
\code{create_curve_table}.
}
\keyword{internal}
