% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputdata.R
\name{create_excel_report}
\alias{create_excel_report}
\title{Create Excel Report}
\usage{
create_excel_report(
  dilution_summary,
  file_name = "dilution_summary.xlsx",
  sheet_name = "DilutionSummary"
)
}
\arguments{
\item{dilution_summary}{The summary table generated by function \code{summarise_dilution_table}}

\item{file_name}{Name of the excel file, Default: 'dilution_summary.xlsx'}

\item{sheet_name}{Sheet name to output the results in Excel, Default: 'DilutionSummary'}
}
\description{
Create Excel Report
}
