% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputdata-ggplot.R
\name{create_ggplot_pdf_report}
\alias{create_ggplot_pdf_report}
\title{Create ggplot dilution plots in a pdf file}
\usage{
create_ggplot_pdf_report(
  ggplot_list,
  filename = "dilution_plot.pdf",
  ncol_per_page = 2,
  nrow_per_page = 2,
  width = 15,
  height = 8,
  testing = FALSE
)
}
\arguments{
\item{ggplot_list}{list of plots to be arranged into a grid
for each pdf page. The list of plot should can be created
via the function \code{create_ggplot_table}. The function will create
a data frame or tibble and the column name "panel" should contain
the list of plots.}

\item{filename}{File name of the pdf document,
Default: 'dilution_plot.pdf'}

\item{ncol_per_page}{Number of columns of plots allowed in one
page of the pdf document
Default: 2}

\item{nrow_per_page}{Number of rows of plots allowed in one
page of the pdf document
Default: 2}

\item{width}{The width of one pdf page, Default: 15}

\item{height}{The height of one pdf page, Default: 8}

\item{testing}{To indicate if we are running a test,
if so, no pdf report is given out
Default: FALSE}
}
\description{
Create a pdf file containing dilution plots in ggplot
}
