% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilplots-ggplot.R
\name{create_ggplot_table}
\alias{create_ggplot_table}
\title{Create a ggplot table}
\usage{
create_ggplot_table(
  dilution_table,
  dilution_summary = NULL,
  grouping_variable = c("Transition_Name", "Dilution_Batch"),
  dil_batch_var = "Dilution_Batch",
  dil_batch_col = c("#377eb8", "#4daf4a", "#9C27B0", "#BCAAA4", "#FF8A65", "#EFBBCF"),
  conc_var = "Dilution_Percent",
  conc_var_units = "\%",
  conc_var_interval = 50,
  signal_var = "Area"
)
}
\arguments{
\item{dilution_table}{Output given from the function \code{create_dilution_table}
It is in long table format with columns indicating at least the
lipid/transition name, the concentration and signal. Other columns may be
present if it is used to group the dilution curve together}

\item{dilution_summary}{The summary table generated
by function \code{summarise_dilution_table} and/or \code{evaluate_linearity}
but it can also be any generic data frame or tibble.
If there is no input given in this, the program will create one using
the function \code{summarise_dilution_table} and \code{evaluate_linearity} with
\code{grouping_variable}, \code{conc_var} and \code{signal_var} as inputs
Default: NULL}

\item{grouping_variable}{A character vector of
column names in \code{dilution_table}to indicate how each dilution curve
should be grouped by,
Default: c("Transition_Name", "Dilution_Batch")}

\item{dil_batch_var}{Column name in \code{dilution_table}
to indicate the group name of each dilution batch,
used to colour the points in the dilution plot
Default: 'Dilution_Batch'}

\item{dil_batch_col}{A vector of colurs to be used for the dilution
batch group named given in \code{dil_batch_var},
Default: c("#377eb8", "#4daf4a", "#9C27B0", "#BCAAA4", "#FF8A65", "#EFBBCF")}

\item{conc_var}{Column name in \code{dilution_table} to indicate concentration
Default: 'Dilution_Percent'}

\item{conc_var_units}{Unit of measure for \code{conc_var}, Default: '\%'}

\item{conc_var_interval}{Distance between two tick labels
in the dilution plot,
Default: 50}

\item{signal_var}{Column name in \code{dilution_table} to indicate signal
Default: 'Area'}
}
\value{
A table with columns from \verb{groupung variable}
and a new column \code{panel} created containing a ggplot dilution plot
in each row. This column is used to create the plot figure in the
pdf report.
}
\description{
Create a ggplot table suited for a pdf report
}
