% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilplots-ggplot.R
\name{create_reg_col_vec}
\alias{create_reg_col_vec}
\title{Create Regression Colour Vector}
\usage{
create_reg_col_vec(
  plot_first_half_lin_reg = FALSE,
  plot_last_half_lin_reg = FALSE
)
}
\arguments{
\item{plot_first_half_lin_reg}{Decide if we plot an extra
regression line that best fits the first half
of \code{conc_var} dilution points.
Default: FALSE}

\item{plot_last_half_lin_reg}{Decide if we plot an extra
regression line that best fits the last half
of \code{conc_var} dilution points.
Default: FALSE}
}
\value{
A named vector in which a linear regression is
named as "Lin" and is given the colour black. A
quadratic regression is named as "Quad" and is given
the colour red. A linear regression of the first
half of the dilution points is named as "Lin First Half"
and is given the colour blue. A linear regression of
the last half of the dilution points is
named as "Lin Last Half" and is given the colour purple.
}
\description{
Internal function that create a named vector
to indicate which regression line has what colour
}
\examples{
# Data Creation
regression_colour_vector <- c(plot_first_half_lin_reg = TRUE,
                              plot_last_half_lin_reg = TRUE)

regression_colour_vector

}
\keyword{internal}
