% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilplots-trellis.R
\name{create_trellis_table}
\alias{create_trellis_table}
\title{Create a trellis table}
\usage{
create_trellis_table(
  dilution_table,
  dilution_summary = NULL,
  cog_df = NULL,
  col_name_vec = "col_name_vec",
  desc_vec = "desc_vec",
  type_vec = "type_vec",
  default_label_vec = "default_label_vec",
  grouping_variable = c("Transition_Name", "Dilution_Batch"),
  sample_name_var = "Sample_Name",
  dil_batch_var = "Dilution_Batch",
  dil_batch_col = c("#377eb8", "#4daf4a", "#9C27B0", "#BCAAA4", "#FF8A65", "#EFBBCF"),
  conc_var = "Dilution_Percent",
  conc_var_units = "\%",
  conc_var_interval = 50,
  signal_var = "Area"
)
}
\arguments{
\item{dilution_table}{Output given from the function \code{create_dilution_table}
It is in long table format with columns indicating at least the
lipid/transition name, the concentration and signal. Other columns may be
present if it is used to group the dilution curve together}

\item{dilution_summary}{The summary table generated
by function \code{summarise_dilution_table} and/or \code{evaluate_linearity}
but it can also be any generic data frame or tibble
Default: NULL}

\item{cog_df}{A data frame or tibble that contains cognostics information
If no input is given the cognostics information generated by function
\code{get_dil_default_cognostics} will be used.
Default: NULL}

\item{col_name_vec}{Column name in \code{cog_df} to indicate the columns
in \code{dilution_summary} that needs to be converted to a cognostics,
Default: 'col_name_vec'}

\item{desc_vec}{Column name in \code{cog_df} to indicate the description
for each cognostics as define in \code{trelliscopejs::cog},
Default: 'desc_vec'}

\item{type_vec}{Column name in \code{cog_df} to indicate the type
of each cognostics as define in \code{trelliscopejs::cog},
Default: 'type_vec'}

\item{default_label_vec}{Column name in \code{cog_df} to indicate the if
the given cognostics is a panel label as define in \code{trelliscopejs::cog},
Default: 'default_label_vec'}

\item{grouping_variable}{A character vector of
column names in \code{dilution_table}to indicate how each dilution curve
should be grouped by. It is also going to be used as a conditional
cognostics in the \code{trelliscopejs} report.
Default: c("Transition_Name", "Dilution_Batch")}

\item{sample_name_var}{Column name in \code{dilution_table}
to indicate the sample name. To be used in the dilution plot
Default: 'Sample_Name'}

\item{dil_batch_var}{Column name in \code{dilution_table}
to indicate the group name of each dilution batch,
used to colour the points in the dilution plot
Default: 'Dilution_Batch'}

\item{dil_batch_col}{A vector of colurs to be used for the dilution
batch group named given in \code{dil_batch_var},
Default: c("#377eb8", "#4daf4a", "#9C27B0", "#BCAAA4", "#FF8A65", "#EFBBCF")}

\item{conc_var}{Column name in \code{dilution_table} to indicate concentration
Default: 'Dilution_Percent'}

\item{conc_var_units}{Unit of measure for \code{conc_var}, Default: '\%'}

\item{conc_var_interval}{Distance between two tick labels in the dilution plot,
Default: 50}

\item{signal_var}{Column name in \code{dilution_table} to indicate signal
Default: 'Area'}
}
\value{
A table that is suited for a \code{trelliscopejs} report with
\verb{groupung variable} columns converted to conditional cognostics,
other columns in \code{dilution_summary} converted to cognostics and
a new column \code{panel} created containing a plotly dilution plot in each row.
This column is used to create the plot figure in the
\code{trelliscopejs} report.
}
\description{
Create a trellis table suited for a \code{trelliscopejs} report
}
