% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilplots-plotly.R
\name{dilution_plot_plotly}
\alias{dilution_plot_plotly}
\title{Plot Dilution Data using plotly}
\usage{
dilution_plot_plotly(
  dilution_data,
  sample_name_var,
  dil_batch_var,
  conc_var,
  conc_var_units,
  conc_var_interval,
  signal_var,
  pal
)
}
\arguments{
\item{dilution_data}{A data frame or tibble containing dilution data}

\item{sample_name_var}{Column name in \code{dilution_data}
to indicate the sample name}

\item{dil_batch_var}{Column name in \code{dilution_data}
to indicate the group name of each dilution batch,
used to colour the points in the dilution plot}

\item{conc_var}{Column name in \code{dilution_data} to indicate concentration}

\item{conc_var_units}{Unit of measure for \code{conc_var} in the dilution plot}

\item{conc_var_interval}{Distance between two tick labels}

\item{signal_var}{Column name in \code{dilution_data} to indicate signal}

\item{pal}{Input palette for each dilution batch group in \code{dil_batch_var}.
It is a named char vector where each value is a colour and
name is a dilution batch group given in \code{dil_batch_var}}
}
\value{
Output plotly dilution plot data of one dilution batch per transition
}
\description{
Plot Dilution Data using plotly
}
