% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilplots-ggplot.R
\name{dilution_summary_group_table}
\alias{dilution_summary_group_table}
\title{Plot Dilution Summary Table For One Group}
\usage{
dilution_summary_group_table(dilution_summary_grp)
}
\arguments{
\item{dilution_summary_grp}{A one row data frame or tibble containing dilution summary}
}
\value{
A \code{gridtable} object consisting of two tables. One from
\code{dilution_summary_char_table} and \code{dilution_summary_num_table}
}
\description{
Plot dilution summary table for one group
}
\examples{
wf1_group <- c("Poor Linearity")
wf2_group <- c("Saturation")
r_corr <- c(0.951956)
pra_linear <- c(65.78711)
mandel_p_val <- c(2.899006e-07)
concavity <- c(-4133.501328)
dilution_summary_grp  <- data.frame(wf1_group = wf1_group,
                                    wf2_group = wf2_group,
                                    r_corr = r_corr,
                                    pra_linear = pra_linear,
                                    mandel_p_val = mandel_p_val,
                                    concavity = concavity)
table <- dilution_summary_group_table(dilution_summary_grp)
grid::grid.draw(table)

# No column case
dilution_summary_grp  <- data.frame()
table <- dilution_summary_group_table(dilution_summary_grp)
table
}
