% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tidy-eval.R
\name{exprs_reduce}
\alias{exprs_reduce}
\title{Expression reduction}
\usage{
exprs_reduce(exprs, op = "+")
}
\arguments{
\item{exprs}{A list of expression output from \code{rlang::exprs}}

\item{op}{Operation used to glue the expression together, Default: '+'.}
}
\value{
A single expression that to be used for
non-standard evaluation like formula in \code{stats::lm()}.
}
\description{
An approach to reducing a list of expressions
into one long expression
}
\details{
Code is taken from
\url{https://community.rstudio.com/t/tidy-evaluation-and-formulae/4561/12}
}
\examples{
exprs <- rlang::exprs(foo(), bar(baz), quux + blip)
exprs_reduce(exprs, "-")
}
