% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognostics.R
\name{get_dil_default_cognostics}
\alias{get_dil_default_cognostics}
\title{Get Dilution Default Cognostics}
\usage{
get_dil_default_cognostics(dilution_summary)
}
\arguments{
\item{dilution_summary}{A data frame or tibble output
from the function \code{get_dilution_summary}}
}
\value{
Output a data frame with added cognostics
relevant to the dilution plot
}
\description{
Function used to add the default cognostics defined in
\code{create_dil_default_cog_df} relevant to the dilution plot
}
