% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputdata-ggplot.R
\name{get_number_of_plots_per_page}
\alias{get_number_of_plots_per_page}
\title{Get the number of plots per pdf page}
\usage{
get_number_of_plots_per_page(ncol = NULL, nrow = NULL)
}
\arguments{
\item{ncol}{Number of columns of plots allowed in one
page of the pdf document, Default: NULL}

\item{nrow}{Number of rows of plots allowed in one
page of the pdf document, Default: NULL}
}
\value{
The number of plots per pdf page as \code{ncol * nrow}
}
\description{
Get the number of plots per pdf page
}
\details{
If both \code{ncol} and \code{nrow} are NULL, return \code{Inf}.
If only \code{nrow} is NULL, return \code{ncol}.
If only \code{ncol} is NULL, return \code{nrow}.
}
\examples{
get_number_of_plots_per_page(ncol = 2,
                             nrow = 2)
}
