% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilplots-plotly.R
\name{plot_curve_plotly}
\alias{plot_curve_plotly}
\title{Plot Dilution Curve Using \code{plotly}}
\usage{
plot_curve_plotly(
  dilution_data,
  title,
  pal,
  sample_name_var = "Sample_Name",
  dil_batch_var = "Dilution_Batch_Name",
  conc_var = "Dilution_Percent",
  conc_var_units = "\%",
  conc_var_interval = 50,
  signal_var = "Area",
  plot_first_half_lin_reg = FALSE,
  plot_last_half_lin_reg = FALSE
)
}
\arguments{
\item{dilution_data}{A data frame or tibble containing dilution data}

\item{title}{Title to use for each dilution plot}

\item{pal}{Input palette for each dilution batch group in \code{dil_batch_var}.
It is a named char vector where each value is a colour and
name is a dilution batch group given in \code{dil_batch_var}}

\item{sample_name_var}{Column name in \code{dilution_data}
to indicate the sample name}

\item{dil_batch_var}{Column name in \code{dilution_data}
to indicate the group name of each dilution batch,
used to colour the points in the dilution plot}

\item{conc_var}{Column name in \code{dilution_data} to indicate concentration}

\item{conc_var_units}{Unit of measure for \code{conc_var} in the dilution plot}

\item{conc_var_interval}{Distance between two tick labels}

\item{signal_var}{Column name in \code{dilution_data} to indicate signal}

\item{plot_first_half_lin_reg}{Decide if we plot an extra regression line
that best fits the first half of \code{conc_var} dilution points.
Default: FALSE}

\item{plot_last_half_lin_reg}{Decide if we plot an extra regression line
that best fits the last half of \code{conc_var} dilution points.
Default: FALSE}
}
\value{
Output \code{plotly} dilution plot data of one dilution batch
per transition
}
\description{
Plot Dilution Data using \code{plotly}
}
\examples{
# Data Creation
dilution_percent <- c(
  10, 20, 25, 40, 50, 60,
  75, 80, 100, 125, 150
)
sample_name <- c(
  "Sample_010a", "Sample_020a",
  "Sample_025a", "Sample_040a", "Sample_050a",
  "Sample_060a", "Sample_075a", "Sample_080a",
  "Sample_100a", "Sample_125a", "Sample_150a"
)
dilution_batch_name <- c(
  "B1", "B1", "B1", "B1", "B1",
  "B1", "B1", "B1", "B1", "B1", "B1"
)
lipid1_area_saturated <- c(
  5748124, 16616414, 21702718, 36191617,
  49324541, 55618266, 66947588, 74964771,
  75438063, 91770737, 94692060
)

dilution_data <- tibble::tibble(
  Sample_Name = sample_name,
  Dilution_Percent = dilution_percent,
  Area = lipid1_area_saturated,
  Dilution_Batch_Name = dilution_batch_name
)
# Get the dilution batch name from dilution_table
dilution_batch_name <- dilution_batch_name \%>\%
  unique() \%>\%
  as.character()

dil_batch_col <- c("#377eb8")

# Create palette for each dilution batch for plotting
pal <- dil_batch_col \%>\%
  stats::setNames(dilution_batch_name)

# Plot the html
p <- plot_curve_plotly(dilution_data,
  title = "Lipid_Saturated",
  sample_name_var = "Sample_Name",
  pal = pal,
  dil_batch_var = "Dilution_Batch_Name",
  conc_var = "Dilution_Percent",
  conc_var_units = "\%",
  conc_var_interval = 50,
  signal_var = "Area"
)
}
