% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilplots-ggplot.R
\name{plot_summary_table}
\alias{plot_summary_table}
\title{Plot Curve Summary Table For One Group}
\usage{
plot_summary_table(curve_summary_grp)
}
\arguments{
\item{curve_summary_grp}{A one row data frame or tibble containing curve summary.}
}
\value{
A \code{gridtable} object consisting of two tables. One from
\code{\link[=plot_summary_table_char]{plot_summary_table_char()}} and \code{\link[=plot_summary_table_num]{plot_summary_table_num()}}.
}
\description{
Plot curve summary table for one group
}
\examples{
wf1_group <- c("Poor Linearity")

wf2_group <- c("Saturation")

r_corr <- c(0.951956)

pra_linear <- c(65.78711)

mandel_p_val <- c(2.899006e-07)

concavity <- c(-4133.501328)

curve_summary_grp <- data.frame(
  wf1_group = wf1_group,
  wf2_group = wf2_group,
  r_corr = r_corr,
  pra_linear = pra_linear,
  mandel_p_val = mandel_p_val,
  concavity = concavity
)

table <- plot_summary_table(curve_summary_grp)

grid::grid.draw(table)

# No column case
curve_summary_grp <- data.frame()

table <- plot_summary_table(curve_summary_grp)

table

}
