% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilplots-ggplot.R
\name{plot_summary_table_char}
\alias{plot_summary_table_char}
\title{Plot Curve Summary Character Table For One Group}
\usage{
plot_summary_table_char(
  curve_summary_grp,
  dilution_summary_grp = lifecycle::deprecated()
)
}
\arguments{
\item{curve_summary_grp}{A one row data frame or tibble containing curve summary}

\item{dilution_summary_grp}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{dilution_summary_grp} was renamed to
\code{curve_summary_grp}.}
}
\value{
A \code{gridtable} object consisting of one table. The first
column is the column names of \code{curve_summary_grp} which
are characters or factors or logical. The second column is their
corresponding values. If there are no character/factor/logical
columns in \code{curve_summary_grp}, NULL will be returned
}
\description{
Plot curve summary character table for one group or batch
}
\examples{
wf1_group <- c("Poor Linearity")

wf2_group <- c("Saturation")

r_corr <- c(0.951956)

pra_linear <- c(65.78711)

mandel_p_val <- c(2.899006e-07)

concavity <- c(-4133.501328)

curve_summary_grp <- data.frame(
  wf1_group = wf1_group,
  wf2_group = wf2_group,
  r_corr = r_corr,
  pra_linear = pra_linear,
  mandel_p_val = mandel_p_val,
  concavity = concavity
)

table <- plot_summary_table_char(curve_summary_grp)

grid::grid.draw(table)

# No character/factor/logical column case
curve_summary_grp <- data.frame(r_corr = r_corr)

table <- plot_summary_table_char(curve_summary_grp)

table

}
