% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilstats.R
\name{summarise_dilution_data}
\alias{summarise_dilution_data}
\title{Summarise Dilution Curve Data For One Group}
\usage{
summarise_dilution_data(dilution_data, conc_var, signal_var, details = FALSE)
}
\description{
Get the summary statistics of the dilution data
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{summarise_dilution_data} was renamed to
\code{summarise_curve_data}.
}
\keyword{internal}
