% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydata.R
\name{summarise_dilution_table}
\alias{summarise_dilution_table}
\title{Summarize Dilution Table}
\usage{
summarise_dilution_table(
  dilution_table,
  grouping_variable = c("Transition_Name", "Dilution_Batch_Name"),
  conc_var = "Dilution_Percent",
  signal_var = "Area"
)
}
\description{
Get the summary statistics of each group from
the dilution table
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{summarise_dilution_table} was renamed to
\code{summarise_curve_table}.
}
\keyword{internal}
