% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tidy-eval.R
\name{syms_reduce}
\alias{syms_reduce}
\title{Symbol reduction}
\usage{
syms_reduce(syms, op = "+")
}
\arguments{
\item{syms}{A vector of characters.}

\item{op}{Operation used to glue the characters together, Default: '+'.}
}
\value{
A single expression that to be used for
non-standard evaluation like formula in \code{stats::lm()}.
}
\description{
An approach to reducing a vector of characters to
a single expression.
}
\details{
From \url{https://github.com/tidyverse/glue/issues/108}, it is not advisable
to generate code expression with strings. This function is meant to provide
a more robust way to generate expressions from strings
Code is taken from
\url{https://community.rstudio.com/t/tidy-evaluation-and-formulae/4561/12}
}
\examples{
syms_reduce(c("a"))
syms_reduce(c("a", "b", "c"))
syms_reduce(c("1()", "`a`e", "_foo"))
syms_reduce(c("a", "b", "c"), op = "*")
syms_reduce(c("a", "b", "c"), op = "plus")
}
