% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputdata.R
\name{two_col_num_cond_format}
\alias{two_col_num_cond_format}
\title{FUNCTION_TITLE}
\usage{
two_col_num_cond_format(
  workbook,
  sheet,
  dilution_summary,
  conditional_column,
  threshold_value,
  pass_criteria = c("above", "below"),
  pass_equality = TRUE
)
}
\arguments{
\item{workbook}{A workbook object from \code{openxlsx}}

\item{sheet}{The name of the sheet to apply the numeric style on \code{workbook}}

\item{dilution_summary}{A data frame or tibble output from the function \code{get_dilution_summary}}

\item{conditional_column}{A character vector to indicate which column in \code{dilution_summary} to use}

\item{threshold_value}{The threshold value to indicate a pass or fail}

\item{pass_criteria}{To indicate pass if the value is above or below threshold value,
Default: c("above", "below")}

\item{pass_equality}{TO indicate if equality to the threshold value is considered a pass or fail,
Default: TRUE}
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
r_corr <- c(0.951956, 0.948683, 0.978057, 0.976462, 0.970618, 0.969348)
pra_linear <- c(65.78711, 64.58687, 90.21257, 89.95473, 72.91220, 72.36528)
mandel_p_val <- c(2.899006e-07, 7.922290e-07, 2.903365e-01, 3.082930e-01,
                  3.195779e-08, 6.366588e-08)
concavity <- c(-4133.501328, -4146.745747, -3.350942, -3.393617,
               0.3942824, 0.4012963)
dilution_summary <- data.frame(r_corr = r_corr, pra_linear = pra_linear,
                               mandel_p_val = mandel_p_val, concavity = concavity)
dilution_summary <- mark_near_zero_columns(dilution_summary)
# Create a new workbook
my_workbook <- openxlsx::createWorkbook()

# Create a new worksheet
openxlsx::addWorksheet(wb = my_workbook, sheetName = "Dilution Summary")

# Write to worksheet as an Excel Table
openxlsx::writeDataTable(wb = my_workbook, sheet = "Dilution Summary",
                         x = dilution_summary,
                         withFilter = TRUE,
                         bandedRows = FALSE
                        )

two_col_num_cond_format(workbook = my_workbook,sheet = "Dilution Summary",
                        dilution_summary = dilution_summary,
                        conditional_column = "r_corr",
                        threshold_value = "0.8",
                        pass_criteria = "above"
                       )
}
