% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognostics.R
\name{update_cogs}
\alias{update_cogs}
\title{Update cognostics}
\usage{
update_cogs(
  dilution_summary,
  cog_df,
  col_name_vec = "col_name_vec",
  desc_vec = "desc_vec",
  type_vec = "type_vec",
  default_label_vec = "default_label_vec"
)
}
\arguments{
\item{dilution_summary}{A data frame or tibble output
from the function \code{get_dilution_summary}}

\item{cog_df}{A data frame or tibble output
from the function \code{create_dil_default_cog_df}}

\item{col_name_vec}{Column name in \code{cog_df} to indicate the columns
in \code{dilution_summary} that needs to be converted to a cognostics,
Default: 'col_name_vec'}

\item{desc_vec}{Column name in \code{cog_df} to indicate the description
for each cognostics as define in \code{trelliscopejs::cog},
Default: 'desc_vec'}

\item{type_vec}{Column name in \code{cog_df} to indicate the type
of each cognostics as define in \code{trelliscopejs::cog},
Default: 'type_vec'}

\item{default_label_vec}{Column name in \code{cog_df} to indicate the if
the given cognostics is a panel label as define in \code{trelliscopejs::cog},
Default: 'default_label_vec'}
}
\value{
A data frame or tibble output with some columns converted
to type cog as defined in \code{trelliscopejs::cog}
}
\description{
Update cognostics
}
\details{
More details in \code{trelliscopejs::cog} can be found in
\url{https://rdrr.io/cran/trelliscopejs/man/cog.html}
}
