% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognostics.R
\name{validate_cog_df}
\alias{validate_cog_df}
\title{Validate Cognostics Data}
\usage{
validate_cog_df(
  cog_df,
  needed_column = c("col_name_vec", "desc_vec", "type_vec", "default_label_vec")
)
}
\arguments{
\item{cog_df}{A data frame or tibble that contains cognostics information}

\item{needed_column}{A vector consisting of needed column names that
must be found in \code{cog_df},
Default: c("col_name_vec", "desc_vec",
"type_vec", "default_label_vec")}
}
\value{
An error if the things in \code{needed_column}
is not found in the Cognostics Data
}
\description{
Validate Cognostics Data
}
