% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydata.R
\name{validate_curve_table}
\alias{validate_curve_table}
\title{Validate Curve Table}
\usage{
validate_curve_table(
  curve_table,
  needed_column = c("Transition_Name", "Dilution_Batch_Name", "Dilution_Percent", "Area")
)
}
\arguments{
\item{curve_table}{A curve table containing data of curve points,
variables to group the curve points together and some additional
annotations about the curve points}

\item{needed_column}{A vector consisting of needed column names that
must be found in \code{curve_table},
Default: c("Transition_Name", "Dilution_Batch_Name",
"Dilution_Percent", "Area")}
}
\value{
An error if the things in \code{needed_column}
is not found in \code{curve_table}
}
\description{
Validate Curve Table
}
\examples{
dilution_percent <- c(
  10, 20, 40, 60, 80, 100,
  10, 20, 40, 60, 80, 100
)
dilution_batch_name <- c(
  "B1", "B1", "B1", "B1", "B1", "B1",
  "B2", "B2", "B2", "B2", "B2", "B2"
)
lipid1_area <- c(
  22561, 31178, 39981, 48390, 52171, 53410,
  32561, 41178, 49981, 58390, 62171, 63410
)
transition_name <- c(
  "Lipid1", "Lipid1", "Lipid1",
  "Lipid1", "Lipid1", "Lipid1",
  "Lipid1", "Lipid1", "Lipid1",
  "Lipid1", "Lipid1", "Lipid1"
)
curve_table <- tibble::tibble(
  Transition_Name = transition_name,
  Dilution_Batch_Name = dilution_batch_name,
  Area = lipid1_area,
  Dilution_Percent = dilution_percent
)
validate_curve_table(
  curve_table = curve_table,
  needed_column = c(
    "Transition_Name",
    "Dilution_Batch_Name",
    "Dilution_Percent",
    "Area"
  )
)

}
