% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilstats.R
\name{validate_dilution_data}
\alias{validate_dilution_data}
\title{Validate Dilution Data}
\usage{
validate_dilution_data(dilution_data, conc_var, signal_var)
}
\description{
Validate Dilution Data
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{validate_dilution_data} was renamed to
\code{validate_curve_data}.
}
\keyword{internal}
