% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydata.R
\name{validate_dilution_table}
\alias{validate_dilution_table}
\title{Validate Dilution Table}
\usage{
validate_dilution_table(
  dilution_table,
  needed_column = c("Transition_Name", "Dilution_Batch_Name", "Dilution_Percent", "Area")
)
}
\description{
Validate Dilution Table
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{validate_dilution_table} was renamed to
\code{validate_curve_table}.
}
\keyword{internal}
