% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydata.R
\name{validate_lipid_data_wide}
\alias{validate_lipid_data_wide}
\title{Validate Lipid Data Wide}
\usage{
validate_lipid_data_wide(lipid_data_wide, needed_column = c("Sample_Name"))
}
\arguments{
\item{lipid_data_wide}{A wide format data frame or tibble
that contains the Sample Name usually at the first column followed
by different Transitions/Lipids. Each Transition/Lipid is meant to provide
a \code{signal_var} for each sample.}

\item{needed_column}{A vector consisting of needed column names that
must be found in \code{lipid_data_wide},
Default: c("Sample_Name")}
}
\value{
An error if the things in \code{needed_column}
is not found in the Lipid Data in wide form
}
\description{
Validate Lipid Data in wide form
}
\examples{
sample_name <- c("Sample_010a", "Sample_020a", "Sample_040a",
                 "Sample_060a", "Sample_080a", "Sample_100a",
                 "Sample_010b", "Sample_020b", "Sample_040b",
                 "Sample_060b", "Sample_080b", "Sample_100b")
lipid1_area <- c(22561, 31178, 39981, 48390, 52171, 53410,
                 32561, 41178, 49981, 58390, 62171, 63410)
lipid2_area <- c(2299075, 4136350, 7020062, 8922063, 9288742, 11365710,
                 2300075, 4137350, 7021062, 8923063, 9289742, 11366710)
lipid_data_wide <- tibble::tibble(Sample_Name = sample_name,
                                  Lipid1 = lipid1_area,
                                  Lipid2 = lipid2_area)
validate_lipid_data_wide(lipid_data_wide,
                         needed_column =c("Sample_Name"))
}
