% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputdata-trellis.R
\name{validate_trellis_table}
\alias{validate_trellis_table}
\title{Validate Trellis Table}
\usage{
validate_trellis_table(
  trellis_table,
  grouping_variable = c("Transition_Name", "Dilution_Batch_Name"),
  panel_variable = "panel"
)
}
\arguments{
\item{trellis_table}{The trellis table generated
by function \code{\link[=add_plotly_panel]{add_plotly_panel()}}}

\item{grouping_variable}{A character vector of
column names in \code{trellis_table} to indicate how each dilution curve
should be grouped by. It is also going to be used as a conditional
cognostics in the \code{trelliscopejs} report. The function will check
if these columns are indeed conditional cognostics
Default: c("Transition_Name", "Dilution_Batch_Name")}

\item{panel_variable}{A column name in \code{trellis_table} to indicate
the list of plots to be used in the \code{trelliscopejs} report. This column
name must be fixed to the name \code{panel} and it must be of class
\code{trelliscope_panels}
Default: 'panel'}
}
\description{
Validate trellis table to see if it can be used to create
\code{trelliscopejs} report
}
\details{
A valid trellis table must have a column named \code{panel} with
a list of plots and class \code{trelliscope_panels}, conditional cognostics
columns as indicated in \code{grouping_variable} and the rest of the columns
as common cognostics
}
\examples{
sample_name <- c(
  "Sample_010a", "Sample_020a", "Sample_040a",
  "Sample_060a", "Sample_080a", "Sample_100a",
  "Sample_010b", "Sample_020b", "Sample_040b",
  "Sample_060b", "Sample_080b", "Sample_100b"
)

lipid1_area <- c(
  22561, 31178, 39981, 48390, 52171, 53410,
  32561, 41178, 49981, 58390, 62171, 63410
)

lipid2_area <- c(
  2299075, 4136350, 7020062, 8922063, 9288742, 11365710,
  2300075, 4137350, 7021062, 8923063, 9289742, 11366710
)

curve_data <- tibble::tibble(
  Sample_Name = sample_name,
  Lipid1 = lipid1_area,
  Lipid2 = lipid2_area
)

dilution_percent <- c(
  10, 20, 40, 60, 80, 100,
  10, 20, 40, 60, 80, 100
)

dilution_batch_name <- c(
  "B1", "B1", "B1", "B1", "B1", "B1",
  "B2", "B2", "B2", "B2", "B2", "B2"
)

curve_batch_annot <- tibble::tibble(
  Sample_Name = sample_name,
  Dilution_Batch_Name = dilution_batch_name,
  Dilution_Percent = dilution_percent
)

# Create dilution table
curve_table <- create_curve_table(
  curve_batch_annot = curve_batch_annot,
  curve_data_wide = curve_data,
  common_column = "Sample_Name",
  signal_var = "Area",
  column_group = "Transition_Name"
)

# Create a trellis table without dilution summary
trellis_table_auto <- add_plotly_panel(
  curve_table,
  grouping_variable = c("Transition_Name",
                        "Dilution_Batch_Name"),
  sample_name_var = "Sample_Name",
  curve_batch_var = "Dilution_Batch_Name",
  conc_var = "Dilution_Percent",
  conc_var_units = "\%",
  conc_var_interval = 50,
  signal_var = "Area"
  ) \%>\%
  convert_to_cog(
    cog_df = NULL,
    grouping_variable = c(
      "Transition_Name",
      "Dilution_Batch_Name"
    ),
    panel_variable = "panel",
    col_name_vec = "col_name_vec",
    desc_vec = "desc_vec",
    type_vec = "type_vec"
  )

# Check if trellis_table_auto is valid
validate_trellis_table(trellis_table_auto)

}
