% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilstats.R
\name{calculate_concavity}
\alias{calculate_concavity}
\title{Calculate Concavity}
\usage{
calculate_concavity(dilution_data, conc_var, signal_var)
}
\arguments{
\item{dilution_data}{A data frame or tibble containing dilution data}

\item{conc_var}{Column name in \code{dilution_data} to indicate concentration}

\item{signal_var}{Column name in \code{dilution_data} to indicate signal}
}
\value{
The concavity of the Dilution Quadratic Model
}
\description{
Calculate the concavity of the Dilution Quadratic Model
}
\details{
The function will return NA if the number of dilution points
is less than or equal to three
}
\examples{
dilution_percent <- c(10, 20, 40, 60, 80, 100)
area <- c(22561, 31178, 39981, 48390, 52171, 53410)
dilution_data <- data.frame(Dilution_Percent = dilution_percent, Area = area)
concavity_value <- calculate_concavity(
  dilution_data,
  "Dilution_Percent", "Area"
)
concavity_value
}
