% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilstats.R
\name{calculate_mandel}
\alias{calculate_mandel}
\title{Calculate Mandel Fitting Test}
\usage{
calculate_mandel(curve_data, conc_var, signal_var)
}
\arguments{
\item{curve_data}{A data frame or tibble containing curve data.}

\item{conc_var}{Column name in \code{curve_data} to indicate concentration.}

\item{signal_var}{Column name in \code{curve_data} to indicate signal.}
}
\value{
A tibble containing Mandel test statistics. The values are
the Mandel test value and its corresponding p value. More information
of the test can be found in
\url{https://link.springer.com/article/10.1007/s00769-002-0487-6}.
}
\description{
Calculate the Mandel fitting test for curve data.
}
\details{
The function will return a tibble with NA values
if the number of curve points is less than or equal to three.
}
\examples{
concentration <- c(10, 20, 40, 60, 80, 100)

signal <- c(22561, 31178, 39981, 48390, 52171, 53410)

curve_data <- data.frame(Concentration = concentration,
                         Signal = signal)

mandel_result <- calculate_mandel(
  curve_data,
  "Concentration",
  "Signal"
)

mandel_result

}
