% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilstats.R
\name{calculate_mandel}
\alias{calculate_mandel}
\title{Calculate the Mandel fitting test}
\usage{
calculate_mandel(dilution_data, conc_var, signal_var)
}
\arguments{
\item{dilution_data}{A data frame or tibble containing dilution data}

\item{conc_var}{Column name in \code{dilution_data} to indicate concentration}

\item{signal_var}{Column name in \code{dilution_data} to indicate signal}
}
\value{
A tibble containing Mandel test statistics. The values are
the Mandel test value and its corresponding p value. More information
of the test can be found in
\url{https://link.springer.com/article/10.1007/s00769-002-0487-6}
}
\description{
Calculate the Mandel fitting test for dilution data
}
\details{
The function will return a tibble with NA values
if the number of dilution points is less than or equal to three.
}
\examples{
dilution_percent <- c(10, 20, 40, 60, 80, 100)
area <- c(22561, 31178, 39981, 48390, 52171, 53410)
dilution_data <- data.frame(Dilution_Percent = dilution_percent, Area = area)
mandel_result <- calculate_mandel(dilution_data, "Dilution_Percent", "Area")
}
