% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilstats.R
\name{calculate_pra_linear}
\alias{calculate_pra_linear}
\title{Calculate Linear Model's Percent Residual Accuracy (PRA)}
\usage{
calculate_pra_linear(dilution_data, conc_var, signal_var)
}
\arguments{
\item{dilution_data}{A data frame or tibble containing dilution data}

\item{conc_var}{Column name in \code{dilution_data} to indicate concentration}

\item{signal_var}{Column name in \code{dilution_data} to indicate signal}
}
\value{
The Percent Residual Accuracy (PRA) of the
Dilution Linear Model. More information of this value can be found in
\url{https://www.sciencedirect.com/science/article/abs/pii/S0039914018307549}
}
\description{
Calculate the Percent Residual Accuracy (PRA) of the
Dilution Linear Model
}
\details{
The function will return NA if the number of dilution points
is less than or equal to three
}
\examples{
dilution_percent <- c(10, 20, 40, 60, 80, 100)
area <- c(22561, 31178, 39981, 48390, 52171, 53410)
dilution_data <- data.frame(Dilution_Percent = dilution_percent, Area = area)
pra_value <- calculate_pra_linear(dilution_data, "Dilution_Percent", "Area")
}
