% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilplots-ggplot.R
\name{dilution_plot_ggplot}
\alias{dilution_plot_ggplot}
\title{Plot Dilution Data using ggplot2}
\usage{
dilution_plot_ggplot(
  dilution_data,
  title,
  dil_batch_var,
  conc_var,
  conc_var_units,
  conc_var_interval,
  signal_var,
  pal
)
}
\arguments{
\item{dilution_data}{A data frame or tibble containing dilution data}

\item{title}{Title to use for each dilution plot}

\item{dil_batch_var}{Column name in \code{dilution_data}
to indicate the group name of each dilution batch,
used to colour the points in the dilution plot}

\item{conc_var}{Column name in \code{dilution_data} to indicate concentration}

\item{conc_var_units}{Unit of measure for \code{conc_var}}

\item{conc_var_interval}{Distance between two tick labels
during \code{conc_var} plotting. Tick label will alway start from 0}

\item{signal_var}{Column name in \code{dilution_data} to indicate signal}

\item{pal}{Input palette for each dilution batch group in \code{dil_batch_var}.
It is a named char vector where each value is a colour and
name is a dilution batch group given in \code{dil_batch_var}}
}
\value{
Output ggplot dilution plot data of one dilution batch per transition
}
\description{
Plot Dilution Data using ggplot2
}
