% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineval.R
\name{evaluate_linearity}
\alias{evaluate_linearity}
\title{Evaluate linearity}
\usage{
evaluate_linearity(
  dilution_summary,
  grouping_variable = c(),
  corrcoef_column = "r_corr",
  corrcoef_min_threshold = 0.8,
  pra_column = "pra_linear",
  pra_min_threshold = 80,
  mandel_p_val_column = "mandel_p_val",
  mandel_p_val_threshold = 0.05,
  concavity_column = "concavity"
)
}
\arguments{
\item{dilution_summary}{A data frame or tibble output from
the function \code{get_dilution_summary}}

\item{grouping_variable}{A character vector of
column names in \code{dilution_summary}to indicate which columns should be placed
first before the evaluation results. Default: c()}

\item{corrcoef_column}{A column in \code{dilution_summary} that holds the
correlation coefficient, Default: 'r_corr'}

\item{corrcoef_min_threshold}{The minimum threshold value of the curve's
correlation coefficient to pass being potentially linear.
Equality to the threshold is considered a pass, Default: 0.8}

\item{pra_column}{A column in \code{dilution_summary} that holds the
percent residual accuracy, Default: 'pra_linear'}

\item{pra_min_threshold}{The minimum threshold value of the curve's
percent residual accuracy to pass being potentially linear.
Equality to the threshold is considered a pass, Default: 80}

\item{mandel_p_val_column}{A column in \code{dilution_summary} that holds the
p value results for the Mandel's fitting test, Default: 'mandel_p_val'}

\item{mandel_p_val_threshold}{The threshold value of the curve's
p value for the Mandel's fitting test to reject the hypothesis that
the quadratic model fits better than the linear model.
Default: 0.05}

\item{concavity_column}{A column in \code{dilution_summary} that holds the
concavity of the quadratic model, Default: 'concavity'}
}
\value{
A data frame or tibble with evaluation results
}
\description{
Evaluate if a curve is linear based
on dilution summary statistics
}
\details{
Two work flows are given to evaluate linearity of dilution
curves. The results are highlighted as columns \code{curve_group1} and
\code{curve_group2} for now. Column names used to categorise the dilution
curves will be moved to the front allow with \code{curve_group1} and
\code{curve_group2}
}
\examples{
r_corr <- c(0.951956, 0.948683, 0.978057, 0.976462, 0.970618, 0.969348)
pra_linear <- c(65.78711, 64.58687, 90.21257, 89.95473, 72.91220, 72.36528)
mandel_p_val <- c(2.899006e-07, 7.922290e-07, 2.903365e-01, 3.082930e-01,
                  3.195779e-08, 6.366588e-08)
concavity <- c(-4133.501328, -4146.745747, -3.350942, -3.393617,
               0.3942824, 0.4012963)
dilution_summary <- data.frame(r_corr = r_corr, pra_linear = pra_linear,
                               mandel_p_val = mandel_p_val,
                               concavity = concavity)
dilution_summary <- evaluate_linearity(dilution_summary)
}
