% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilplots-plotly.R
\name{plot_curve_plotly}
\alias{plot_curve_plotly}
\title{Plot Curve Using \code{plotly}}
\usage{
plot_curve_plotly(
  curve_data,
  title = "",
  pal,
  sample_name_var = "Sample_Name",
  curv_batch_var = "Dilution_Batch_Name",
  dilution_data = lifecycle::deprecated(),
  dil_batch_var = lifecycle::deprecated(),
  conc_var = "Dilution_Percent",
  conc_var_units = "\%",
  conc_var_interval = 50,
  signal_var = "Area",
  plot_first_half_lin_reg = FALSE,
  plot_last_half_lin_reg = FALSE
)
}
\arguments{
\item{curve_data}{A data frame or tibble containing curve data.}

\item{title}{Title to use for each curve plot.}

\item{pal}{Input palette for each curve batch group in \code{curv_batch_var}.
It is a named char vector where each value is a colour and
name is a curve batch group given in \code{curv_batch_var}.}

\item{sample_name_var}{Column name in \code{curve_data}
to indicate the sample name.}

\item{curv_batch_var}{Column name in \code{curve_data}
to indicate the group name of each curve batch,
used to colour the points in the curve plot.}

\item{dilution_data}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{dilution_data} was renamed to
\code{curve_data}.}

\item{dil_batch_var}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{dil_batch_var} was renamed to
\code{curv_batch_var}.}

\item{conc_var}{Column name in \code{curve_data} to indicate concentration.}

\item{conc_var_units}{Unit of measure for \code{conc_var} in the curve plot.}

\item{conc_var_interval}{Distance between two tick labels.}

\item{signal_var}{Column name in \code{curve_data} to indicate signal.}

\item{plot_first_half_lin_reg}{Decide if we plot an extra regression line
that best fits the first half of \code{conc_var} curve points.
Default: FALSE}

\item{plot_last_half_lin_reg}{Decide if we plot an extra regression line
that best fits the last half of \code{conc_var} curve points.
Default: FALSE}
}
\value{
Output \code{plotly} curve plot data of one curve batch
per transition.
}
\description{
Plot curve using \code{plotly}.
}
\examples{

# Data Creation
dilution_percent <- c(
  10, 20, 25, 40, 50, 60,
  75, 80, 100, 125, 150
)

sample_name <- c(
  "Sample_010a", "Sample_020a",
  "Sample_025a", "Sample_040a", "Sample_050a",
  "Sample_060a", "Sample_075a", "Sample_080a",
  "Sample_100a", "Sample_125a", "Sample_150a"
)

dilution_batch_name <- c(
  "B1", "B1", "B1", "B1", "B1",
  "B1", "B1", "B1", "B1", "B1", "B1"
)

lipid1_area_saturated <- c(
  5748124, 16616414, 21702718, 36191617,
  49324541, 55618266, 66947588, 74964771,
  75438063, 91770737, 94692060
)

curve_data <- tibble::tibble(
  Sample_Name = sample_name,
  Dilution_Percent = dilution_percent,
  Area = lipid1_area_saturated,
  Dilution_Batch_Name = dilution_batch_name
)

# Get the curve batch name from curve_table
curve_batch_name <- dilution_batch_name \%>\%
  unique() \%>\%
  as.character()

curv_batch_col <- c("#377eb8")

# Create palette for each curve batch for plotting
pal <- curv_batch_col \%>\%
  stats::setNames(curve_batch_name)

# Plot the html
p <- plot_curve_plotly(curve_data,
  title = "Lipid_Saturated",
  sample_name_var = "Sample_Name",
  pal = pal,
  curv_batch_var = "Dilution_Batch_Name",
  conc_var = "Dilution_Percent",
  conc_var_units = "\%",
  conc_var_interval = 50,
  signal_var = "Area"
)

p

}
