% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognostics.R
\name{update_cog_auto}
\alias{update_cog_auto}
\title{Update Cognostics Automatically}
\usage{
update_cog_auto(dilution_summary)
}
\arguments{
\item{dilution_summary}{The summary table generated
by function \code{\link[=summarise_dilution_table]{summarise_dilution_table()}} and/or
\code{\link[=evaluate_linearity]{evaluate_linearity()}}}
}
\value{
\code{dilution_summary} with some columns converted
to type cog as defined in \code{trelliscopejs::cog}
}
\description{
Update cognostics on the dilution summary based on
the cognostics data frame given by the function \code{create_default_cog_df}.
}
\details{
More details in \code{trelliscopejs::cog} can be found in
\url{https://rdrr.io/cran/trelliscopejs/man/cog.html}
}
\examples{
dilution_percent <- c(
  10, 20, 25, 40, 50, 60,
  75, 80, 100, 125, 150,
  10, 25, 40, 50, 60,
  75, 80, 100, 125, 150
)
dilution_batch_name <- c(
  "B1", "B1", "B1", "B1", "B1",
  "B1", "B1", "B1", "B1", "B1", "B1",
  "B2", "B2", "B2", "B2", "B2",
  "B2", "B2", "B2", "B2", "B2"
)
sample_name <- c(
  "Sample_010a", "Sample_020a",
  "Sample_025a", "Sample_040a", "Sample_050a",
  "Sample_060a", "Sample_075a", "Sample_080a",
  "Sample_100a", "Sample_125a", "Sample_150a",
  "Sample_010b", "Sample_025b",
  "Sample_040b", "Sample_050b", "Sample_060b",
  "Sample_075b", "Sample_080b", "Sample_100b",
  "Sample_125b", "Sample_150b"
)
lipid1_area_saturated <- c(
  5748124, 16616414, 21702718, 36191617,
  49324541, 55618266, 66947588, 74964771,
  75438063, 91770737, 94692060,
  5192648, 16594991, 32507833, 46499896,
  55388856, 62505210, 62778078, 72158161,
  78044338, 86158414
)
lipid2_area_linear <- c(
  31538, 53709, 69990, 101977, 146436, 180960,
  232881, 283780, 298289, 344519, 430432,
  25463, 63387, 90624, 131274, 138069,
  205353, 202407, 260205, 292257, 367924
)
lipid3_area_lod <- c(
  544, 397, 829, 1437, 1808, 2231,
  3343, 2915, 5268, 8031, 11045,
  500, 903, 1267, 2031, 2100,
  3563, 4500, 5300, 8500, 10430
)
lipid4_area_nonlinear <- c(
  380519, 485372, 478770, 474467, 531640, 576301,
  501068, 550201, 515110, 499543, 474745,
  197417, 322846, 478398, 423174, 418577,
  426089, 413292, 450190, 415309, 457618
)

curve_annot <- tibble::tibble(
  Sample_Name = sample_name,
  Dilution_Batch_Name = dilution_batch_name,
  Dilution_Percent = dilution_percent
)
curve_data <- tibble::tibble(
  Sample_Name = sample_name,
  Lipid1 = lipid1_area_saturated,
  Lipid2 = lipid2_area_linear,
  Lipid3 = lipid3_area_lod,
  Lipid4 = lipid4_area_nonlinear
)


# Create dilution table
curve_table <- create_curve_table(
  curve_annot = curve_annot,
  curve_data_wide = curve_data,
  common_column = "Sample_Name",
  signal_var = "Area",
  column_group = "Transition_Name"
)

# Create dilution statistical summary
dilution_summary <- curve_table \%>\%
  summarise_dilution_table(
    grouping_variable = c(
      "Transition_Name",
      "Dilution_Batch_Name"
    ),
    conc_var = "Dilution_Percent",
    signal_var = "Area"
  ) \%>\%
  dplyr::arrange(.data$Transition_Name) \%>\%
  evaluate_linearity(grouping_variable = c(
    "Transition_Name",
    "Dilution_Batch_Name"
  ))

updated_summary <- update_cog_auto(dilution_summary)

# Observe that the columns has been converted
# to class cognostics
dilution_summary
updated_summary
}
