% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognostics.R
\name{validate_cog_df}
\alias{validate_cog_df}
\title{Validate Cognostics Data}
\usage{
validate_cog_df(
  cog_df,
  needed_column = c("col_name_vec", "desc_vec", "type_vec")
)
}
\arguments{
\item{cog_df}{A data frame or tibble that contains cognostics information}

\item{needed_column}{A vector consisting of needed column names that
must be found in \code{cog_df},
Default:
\code{c("col_name_vec", "desc_vec", "type_vec")}}
}
\value{
An error if the things in \code{needed_column}
is not found in the Cognostics Data
}
\description{
Validate Cognostics Data
}
\examples{
# Create Cognostics Dataframe
col_name_vec <- c("Curve_Name", "Curve_Batch_Name")

desc_vec <- c("Curve_Name", "Curve_Batch_Name")

type_vec <- c("factor", "factor")

cog_df <- data.frame(
  col_name_vec = col_name_vec,
  desc_vec = desc_vec,
  type_vec = type_vec
)

validate_cog_df(cog_df)
}
