% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilstats.R
\name{validate_curve_data}
\alias{validate_curve_data}
\title{Validate Curve Data}
\usage{
validate_curve_data(curve_data, conc_var, signal_var)
}
\arguments{
\item{curve_data}{A data frame or tibble containing curve data.}

\item{conc_var}{Column name in \code{curve_data} to indicate concentration.}

\item{signal_var}{Column name in \code{curve_data} to indicate signal.}
}
\value{
An error if the column name is not found in the \code{curve_data}.
}
\description{
Validate Curve Data.
}
\examples{
concentration <- c(10, 20, 40, 60, 80, 100)

signal <- c(22561, 31178, 39981, 48390, 52171, 53410)

curve_data <- data.frame(Concentration = concentration,
                         Signal = signal)

validate_curve_data(
  curve_data = curve_data,
  conc_var = "Concentration",
  signal_var = "Signal"
)

}
