% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydata.R
\name{validate_dilution_annot}
\alias{validate_dilution_annot}
\title{Validate Dilution Annotation}
\usage{
validate_dilution_annot(dilution_annot, needed_column = c("Sample_Name"))
}
\description{
Validate Dilution Annotation
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{validate_dilution_annot} was renamed to
\code{validate_curve_batch_annot}.
}
\keyword{internal}
