% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilstats.R
\name{validate_dilution_data}
\alias{validate_dilution_data}
\title{Calculate the Percent Residual Accuracy (PRA)}
\usage{
validate_dilution_data(dilution_data, conc_var, signal_var, validator)
}
\arguments{
\item{dilution_data}{A data frame or tibble containing dilution data}

\item{conc_var}{Column name in \code{diltuion_data} to indicate concentration}

\item{signal_var}{Column name in \code{diltuion_data} to indicate signal}

\item{validator}{An R object created by \code{data.validator::create_validator()}}
}
\value{
An updated \code{data.validator::create_validator()} R object to be used
to generate a reports based on \code{assertr}'s validation
}
\description{
Calculate the Percent Residual Accuracy (PRA) of the
Dilution Linear Model
}
\details{
More information on the R package \code{data.validator} can be found in
\url{https://github.com/Appsilon/data.validator}
}
\examples{
dilution_percent <- c(10, 20, 40, 60, 80, 100)
area <- c(22561, 31178, 39981, 48390, 52171, 53410)
dilution_data <- data.frame(Dilution_Percent = dilution_percent, Area = area)
validator <- data.validator::create_validator()
validator <- validate_dilution_data(dilution_data,
                                   "Dilution_Percen", "Are",
                                   validator)
}
