% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydata.R
\name{validate_lipid_data_wide}
\alias{validate_lipid_data_wide}
\title{Validate Lipid Data Wide}
\usage{
validate_lipid_data_wide(lipid_data_wide, needed_column = c("Sample_Name"))
}
\description{
Validate Lipid Data in wide form
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{validate_lipid_data_wide} was renamed to
\code{validate_curve_data_wide}.
}
\keyword{internal}
